Remove updateSectionInfo hook
It turns out that that method is irrelevant for the purposes of the
tweak. That method is in fact the source of the bug that allows one to
enable MAP notifications for an arbitrary section simply by flipping it
on and off in Settings. I am not sure if that is a bug or a feature...